add utils assay
