+++ celescope/__init__.py
@@ -19,7 +19,8 @@ ASSAY_LIST = [
     'flv_CR',
     'convert10X',
     'flv_trust4',
-    'rna_virus'
+    'rna_virus',
+    'utils'
 ]
 
 ROOT_PATH = os.path.dirname(__file__)
