+++ celescope/celescope.py
@@ -1,14 +1,8 @@
 import argparse
-import os
-import celescope
-import subprocess
 
 from celescope.tools import utils
 from celescope.__init__ import __VERSION__, ASSAY_LIST
 
-TOOLS_DIR = os.path.dirname(celescope.tools.__file__)
-SINGLE_MODULE = ['mkgtf']
-ASSAY_LIST += SINGLE_MODULE
 
 class ArgFormatter(argparse.ArgumentDefaultsHelpFormatter, argparse.RawTextHelpFormatter):
     pass
@@ -24,15 +18,6 @@ def main():
     for assay in ASSAY_LIST:
         text = utils.get_assay_text(assay)
         subparser_1st = subparsers.add_parser(assay, description=text)
-
-        if assay == 'mkgtf':
-            def func(args):
-                subprocess.check_call(f'python {TOOLS_DIR}/mkgtf.py {args.gtf} {args.filtered_gtf}', shell=True)
-            
-            subparser_1st.add_argument('gtf', help='raw gtf file')
-            subparser_1st.add_argument('filtered_gtf', help='filtered gtf file')
-            subparser_1st.set_defaults(func=func)
-            continue
 
         # add 2ed subparser
         subparser_2nd = subparser_1st.add_subparsers()
