+++ celescope/utils/__init__.py
@@ -1,0 +1,2 @@
+__STEPS__ = ['mkgtf']
+__ASSAY__ = 'utils'
