+++ docs/quick_start.md
@@ -39,10 +39,11 @@ gunzip Homo_sapiens.GRCh38.dna.primary_assembly.fa.gz
 gunzip Homo_sapiens.GRCh38.99.gtf.gz
 
 conda activate celescope
+celescope utils mkgtf Homo_sapiens.GRCh38.99.gtf Homo_sapiens.GRCh38.99.filtered.gtf
 celescope rna mkref \
  --genome_name Homo_sapiens_ensembl_99 \
  --fasta Homo_sapiens.GRCh38.dna.primary_assembly.fa \
- --gtf Homo_sapiens.GRCh38.99.gtf
+ --gtf Homo_sapiens.GRCh38.99.filtered.gtf
 ```
 
 ### Mus musculus
@@ -58,10 +59,12 @@ gunzip Mus_musculus.GRCm38.dna.primary_assembly.fa.gz
 gunzip Mus_musculus.GRCm38.99.gtf.gz
 
 conda activate celescope
+celescope utils mkgtf Mus_musculus.GRCm38.99.gtf Mus_musculus.GRCm38.99.filtered.gtf
+
 celescope rna mkref \
  --genome_name Mus_musculus_ensembl_99 \
  --fasta Mus_musculus.GRCm38.dna.primary_assembly.fa \
- --gtf Mus_musculus.GRCm38.99.gtf
+ --gtf Mus_musculus.GRCm38.99.filtered.gtf
 ```
 
 2. Generate scripts for each sample
